resume-parser-service/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
pdfplumber
python-dotenv
google-generativeai
